# model/field: avoid closures