--- a/model/field/rum_v3_mapping.go
+++ b/model/field/rum_v3_mapping.go
@@ -121,30 +121,42 @@ var rumV3Mapping = map[string]string{
 	"version":                     "ve",
 }
 
-func Mapper(shortFieldNames bool) func(string) string {
-	return func(s string) string {
-		shortField, ok := rumV3Mapping[s]
-		if ok && shortFieldNames {
-			return shortField
-		}
-		return s
+var rumV3InverseMapping = make(map[string]string)
+
+func init() {
+	for k, v := range rumV3Mapping {
+		rumV3InverseMapping[v] = k
 	}
 }
 
-var rumV3InverseMapping = func() map[string]string {
-	m := make(map[string]string)
-	for k, v := range rumV3Mapping {
-		m[v] = k
+func Mapper(shortFieldNames bool) func(string) string {
+	if shortFieldNames {
+		return rumV3Mapper
 	}
-	return m
-}()
+	return identityMapper
+}
 
 func InverseMapper(shortFieldNames bool) func(string) string {
-	return func(s string) string {
-		longField, ok := rumV3InverseMapping[s]
-		if ok && shortFieldNames {
-			return longField
-		}
-		return s
+	if shortFieldNames {
+		return rumV3InverseMapper
 	}
+	return identityMapper
+}
+
+func rumV3Mapper(long string) string {
+	if short, ok := rumV3Mapping[long]; ok {
+		return short
+	}
+	return long
+}
+
+func rumV3InverseMapper(short string) string {
+	if long, ok := rumV3InverseMapping[short]; ok {
+		return long
+	}
+	return short
+}
+
+func identityMapper(s string) string {
+	return s
 }